Refactored ThreeSubscriptions to take bundles as props, and created SubscriptionsSection for showcase page.
