+++ assets/components/threeSubscriptions/gridImageProperties.js
@@ -1,0 +1,14 @@
+// @flow
+
+// ----- Properties ----- //
+
+const gridImageProperties = {
+  srcSizes: [825, 500, 140],
+  sizes: '(max-width: 660px) 165px, (max-width: 740px) 174px, (max-width: 980px) 196px, (max-width: 1140px) 205px, 165px',
+  imgType: 'png',
+};
+
+
+// ----- Exports ----- //
+
+export { gridImageProperties };
